move about home and activity to folder View and erase 404
