--- src/components/Error404/error404.css
@@ -1,15 +1,0 @@
-.color_red {
-    color: #db5752;
-}
-
-.color_yellow {
-    color: #fafa88;
-}
-
-.color_blue {
-    color: #9AC9FB;
-}
-
-.font_size {
-    font-size: 250px;
-}
