+++ src/components/View/Activities/activity.css
@@ -20,6 +20,15 @@
   background-color: rgba(189, 189, 189, 0.472);
 }
 
+.container-error {
+  border: 1px solid #18a0fb;
+  height: 400px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(189, 189, 189, 0.472);
+}
+
 @media only screen and (max-width: 896px) {
   .button_style {
     width: 95%;
