+++ src/routes/AppRoutes.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 
-import Home from '../components/Home/Home.jsx';
-import About from '../components/About/About.jsx';
-import ActivityView from '../components/Activities/ActivityView';
+import Home from '../components/View/Home/Home';
+import About from '../components/View/About/About.jsx';
+import Activity from '../components/View/Activities/Activity';
 
 export default function AppRoutes() {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
       <Route path="/about" element={<About />} />
-      <Route path="/actividades/:id" element={<ActivityView />}/>
+      <Route path="/actividades/:id" element={<Activity />}/>
     </Routes>
   )
 }
